# gen/api/openapi.yaml
openapi: 3.0.1
info:
  title: Main service API
  version: "1.0"
servers:
- description: Generated server url
  url: http://localhost:8080
...
paths:
...
  /admin/compilations/{compId}/events/{eventId}:
    delete:
      operationId: removeEventFromCompilation
      parameters:
      - description: id подборки
        explode: false
        in: path
        name: compId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: id события
        explode: false
        in: path
        name: eventId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Событие удалено из подборки
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Удалить событие из подборки
      tags:
      - 'Admin: Подборки событий'
      x-accepts: application/json
    patch:
      operationId: addEventToCompilation
      parameters:
      - description: id подборки
        explode: false
        in: path
        name: compId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: id события
        explode: false
        in: path
        name: eventId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          description: Событие добавлено
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Добавить событие в подборку
      tags:
      - 'Admin: Подборки событий'
      x-accepts: application/json
...
  /admin/events:
    get:
      description: Эндпоинт возвращает полную информацию обо всех событиях подходящих
        под переданные условия
      operationId: getEvents_2
      parameters:
      - description: список id пользователей, чьи события нужно найти
        explode: true
        in: query
        name: users
        required: false
        schema:
          items:
            format: int64
            type: integer
          type: array
        style: form
      - description: список состояний в которых находятся искомые события
        explode: true
        in: query
        name: states
        required: false
        schema:
          items:
            type: string
          type: array
        style: form
      - description: список id категорий в которых будет вестись поиск
        explode: true
        in: query
        name: categories
        required: false
        schema:
          items:
            format: int64
            type: integer
          type: array
        style: form
      - description: дата и время не раньше которых должно произойти событие
        explode: true
        in: query
        name: rangeStart
        required: false
        schema:
          type: string
        style: form
      - description: дата и время не позже которых должно произойти событие
        explode: true
        in: query
        name: rangeEnd
        required: false
        schema:
          type: string
        style: form
      - description: количество событий, которые нужно пропустить для формирования
          текущего набора
        explode: true
        in: query
        name: from
        required: false
        schema:
          default: 0
          format: int32
          type: integer
        style: form
      - description: количество событий в наборе
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventFullDto'
                type: array
          description: События найдены
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Поиск событий
      tags:
      - 'Admin: События'
      x-accepts: application/json
  /admin/events/{eventId}:
    put:
      description: Редактирование данных любого события администратором. Валидация
        данных не требуется.
      operationId: updateEvent
      parameters:
      - description: id события
        explode: false
        in: path
        name: eventId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateEventRequest'
        description: Данные для изменения информации о событии
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие отредактировано
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Редактирование события
      tags:
      - 'Admin: События'
      x-contentType: application/json
      x-accepts: application/json
...
  /events:
    get:
      description: "Обратите внимание: \n- это публичный эндпоинт, соответственно\
        \ в выдаче должны быть только опубликованные события\n- текстовый поиск (п\
        о аннотации и подробному описанию) должен быть без учета регистра букв\n-\
        \ если в запросе не указан диапазон дат [rangeStart-rangeEnd], то нужно вы\
        гружать события, которые произойдут позже текущей даты и времени\n- информ\
        ация о каждом событии должна включать в себя количество просмотров и колич\
        ество уже одобренных заявок на участие\n- информацию о том, что по этому э\
        ндпоинту был осуществлен и обработан запрос, нужно сохранить в сервисе ста\
        тистики"
      operationId: getEvents_1
      parameters:
      - description: текст для поиска в содержимом аннотации и подробном описании
          события
        explode: true
        in: query
        name: text
        required: false
        schema:
          type: string
        style: form
      - description: список идентификаторов категорий в которых будет вестись поиск
        explode: true
        in: query
        name: categories
        required: false
        schema:
          items:
            format: int64
            type: integer
          type: array
        style: form
      - description: поиск только платных/бесплатных событий
        explode: true
        in: query
        name: paid
        required: false
        schema:
          type: boolean
        style: form
      - description: дата и время не раньше которых должно произойти событие
        explode: true
        in: query
        name: rangeStart
        required: false
        schema:
          type: string
        style: form
      - description: дата и время не позже которых должно произойти событие
        explode: true
        in: query
        name: rangeEnd
        required: false
        schema:
          type: string
        style: form
      - description: только события у которых не исчерпан лимит запросов на участие
        explode: true
        in: query
        name: onlyAvailable
        required: false
        schema:
          default: false
          type: boolean
        style: form
      - description: 'Вариант сортировки: по дате события или по количеству просмотров'
        explode: true
        in: query
        name: sort
        required: false
        schema:
          enum:
          - EVENT_DATE
          - VIEWS
          type: string
        style: form
      - description: количество событий, которые нужно пропустить для формирования
          текущего набора
        explode: true
        in: query
        name: from
        required: false
        schema:
          default: 0
          format: int32
          type: integer
        style: form
      - description: количество событий в наборе
        explode: true
        in: query
        name: size
        required: false
        schema:
          default: 10
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/EventShortDto'
                type: array
          description: События найдены
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Получение событий с возможностью фильтрации
      tags:
      - 'Public: События'
      x-accepts: application/json
  /events/{id}:
    get:
      description: |-
        Обратите внимание:
        - событие должно быть опубликовано
        - информация о событии должна включать в себя количество просмотров и количество подтвержденных запросов
        - информацию о том, что по этому эндпоинту был осуществлен и обработан запрос, нужно сохранить в сервисе статистики
      operationId: getEvent_1
      parameters:
      - description: id события
        explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие найдено
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Получение подробной информации об опубликованном событии по его идентификатору
      tags:
      - 'Public: События'
      x-accepts: application/json
...
  /users/{userId}/events/{eventId}:
    get:
      operationId: getEvent
      parameters:
      - description: id текущего пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: id события
        explode: false
        in: path
        name: eventId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие найдено
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Получение полной информации о событии добавленном текущим пользователем
      tags:
      - 'Private: События'
      x-accepts: application/json
    patch:
      description: 'Обратите внимание: Отменить можно только событие в состоянии ожидания
        модерации.'
      operationId: cancelEvent
      parameters:
      - description: id текущего пользователя
        explode: false
        in: path
        name: userId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      - description: id отменяемого события
        explode: false
        in: path
        name: eventId
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EventFullDto'
          description: Событие обновлено
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос составлен с ошибкой
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Не выполнены условия для совершения операции
        "404":
          content:
            application/json:
              example:
                status: NOT_FOUND
                reason: The required object was not found.
                message: Event with id=21 was not found.
                timestamp: 2022-09-07 09:10:50
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Объект не найден
        "409":
          content:
            application/json:
              example:
                status: CONFLICT
                reason: Integrity constraint has been violated
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:47:58
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Запрос приводит к нарушению целостности данных
        "500":
          content:
            application/json:
              example:
                status: INTERNAL_SERVER_ERROR
                reason: Error occurred
                message: 'could not execute statement; SQL [n/a]; constraint [uq_category_name];
                  nested exception is org.hibernate.exception.ConstraintViolationException:
                  could not execute statement'
                timestamp: 2022-09-07 10:23:27
              schema:
                $ref: '#/components/schemas/ApiError'
          description: Внутренняя ошибка сервера
      summary: Отмена события добавленного текущим пользователем.
      tags:
      - 'Private: События'
      x-accepts: application/json
...
components:
  schemas:
    AdminUpdateEventRequest:
      description: Информация для редактирования события администратором. Все поля
        необязательные. Значение полей не валидируется.
      example:
        annotation: Новое краткое описание
        paid: false
        description: Новое полное описание
        requestModeration: false
        location:
          lon: 37.62
          lat: 55.754167
        category: 0
        title: Новый заголовок
        participantLimit: 33
        eventDate: 2025-01-01 09:08:07
      properties:
        annotation:
          description: Краткое описание события
          example: Новое краткое описание
          type: string
        category:
          description: id категории к которой относится событие
          format: int64
          type: integer
        description:
          description: Полное описание события
          example: Новое полное описание
          type: string
        eventDate:
          description: Дата и время на которые намечено событие (в формате "yyyy-MM-dd
            HH:mm:ss")
          example: 2025-01-01 09:08:07
          type: string
        location:
          $ref: '#/components/schemas/Location'
        paid:
          description: Нужно ли оплачивать участие в событии
          example: false
          type: boolean
        participantLimit:
          description: Ограничение на количество участников. Значение 0 - означает
            отсутствие ограничения
          example: 33
          format: int32
          type: integer
        requestModeration:
          description: Нужна ли пре-модерация заявок на участие
          example: false
          type: boolean
        title:
          description: Заголовок события
          example: Новый заголовок
          type: string
      type: object
    ApiError:
      description: Сведения об ошибке
      properties:
        errors:
          description: Список стектрейсов или описания ошибок
          example: []
          items:
            description: Список стектрейсов или описания ошибок
            example: '[]'
            type: string
          type: array
        message:
          description: Сообщение об ошибке
          example: Only pending or canceled events can be changed
          type: string
        reason:
          description: Общее описание причины ошибки
          example: For the requested operation the conditions are not met.
          type: string
        status:
          description: Код статуса HTTP-ответа
          enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
          example: FORBIDDEN
          type: string
        timestamp:
          description: Дата и время когда произошла ошибка (в формате "yyyy-MM-dd
            HH:mm:ss")
          example: 2022-06-09 06:27:23
          type: string
      type: object
...
    Location:
      description: Широта и долгота места проведения события
      example:
        lon: 37.62
        lat: 55.754167
      properties:
        lat:
          description: Широта
          example: 55.754167
          format: float
          type: number
        lon:
          description: Долгота
          example: 37.62
          format: float
          type: number
      type: object
...